+++ Crud/src/main/java/com/KnowIt/Gym_intellect_Crud/Repository/DietMealRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.KnowIt.Gym_intellect_Crud.Entity.DietMeal;
 
 @Repository
-public interface DietMealRepository extends JpaRepository<DietMeal,Integer>{
+public interface DietMealRepository extends JpaRepository<DietMeal,Long>{
 
 }
