+++ Crud/src/main/java/com/KnowIt/Gym_intellect_Crud/Repository/DietPlanTypeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.KnowIt.Gym_intellect_Crud.Entity.DietPlanType;
 
 @Repository
-public interface DietPlanTypeRepository extends JpaRepository<DietPlanType, Integer> {
+public interface DietPlanTypeRepository extends JpaRepository<DietPlanType, Long> {
 }
